+++ src/test/java/com/revature/stepsimplementation/CheckoutHooks.java
@@ -1,33 +1,37 @@
-package com.revature.stepsImplementation;
+package com.revature.stepsimplementation;
 
-import com.revature.pages.FrontPage;
-import com.revature.pages.LoginPage;
-import com.revature.pages.OrdersPage;
-import com.revature.pages.ProductDetailsPage;
+import com.revature.pages.*;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class OrderHistoryHooks {
+import java.time.Duration;
+
+public class CheckoutHooks {
     public static WebDriver driver;
     public static WebDriverWait wait;
+    public static Actions actions;
     public static LoginPage loginPage;
     public static FrontPage frontPage;
-    public static OrdersPage orderPage;
-    public static ProductDetailsPage productDetailsPage;
+    public static CartPage cartPage;
+    public static CheckoutPage checkoutPage;
+
 
     @Before
     public void setUp() {
         WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver();
-        wait = new WebDriverWait(driver, 10);
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        actions = new Actions(driver);
         loginPage = new LoginPage(driver);
         frontPage = new FrontPage(driver);
-        orderPage = new OrdersPage(driver);
-        productDetailsPage = new ProductDetailsPage(driver);
+        cartPage = new CartPage(driver);
+        checkoutPage = new CheckoutPage(driver);
+
     }
 
     @After
