+++ src/test/java/com/revature/stepsimplementation/OrderHistoryHooks.java
@@ -1,6 +1,9 @@
-package com.revature.stepsImplementation;
+package com.revature.stepsimplementation;
 
-import com.revature.pages.*;
+import com.revature.pages.FrontPage;
+import com.revature.pages.LoginPage;
+import com.revature.pages.OrdersPage;
+import com.revature.pages.ProductDetailsPage;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.github.bonigarcia.wdm.WebDriverManager;
@@ -8,25 +11,25 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class CheckoutHooks {
+import java.time.Duration;
+
+public class OrderHistoryHooks {
     public static WebDriver driver;
     public static WebDriverWait wait;
     public static LoginPage loginPage;
     public static FrontPage frontPage;
-    public static CartPage cartPage;
-    public static CheckoutPage checkoutPage;
-
+    public static OrdersPage orderPage;
+    public static ProductDetailsPage productDetailsPage;
 
     @Before
     public void setUp() {
         WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver();
-        wait = new WebDriverWait(driver, 10);
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
         loginPage = new LoginPage(driver);
         frontPage = new FrontPage(driver);
-        cartPage = new CartPage(driver);
-        checkoutPage = new CheckoutPage(driver);
-
+        orderPage = new OrdersPage(driver);
+        productDetailsPage = new ProductDetailsPage(driver);
     }
 
     @After
